+++ pkg/bpv7/extension_block_prophet.go
@@ -1,5 +1,5 @@
 // SPDX-FileCopyrightText: 2019, 2021 Markus Sommer
-// SPDX-FileCopyrightText: 2020 Alvar Penning
+// SPDX-FileCopyrightText: 2020, 2022 Alvar Penning
 //
 // SPDX-License-Identifier: GPL-3.0-or-later
 
@@ -41,6 +41,10 @@ func (pBlock *ProphetBlock) BlockTypeName() string {
 }
 
 func (pBlock ProphetBlock) CheckValid() error {
+	return nil
+}
+
+func (pBlock ProphetBlock) CheckContextValid(*Bundle) error {
 	return nil
 }